Some fixes for windows platform (#9)
* fix Terminal.cpp missing include on windows

* fix Terminal.hpp missing Handle definition on windows

* fix constexpr variables and color definition for TextColor.cpp

* fix constexpr variables and color definition for BackgroundColor.cpp

* hide implementation details of TextModifier class

* fix unwanted narrowing in Terminal.cpp

* fix windows.h include when build on non-windows platform

+++ Terminal/src/Terminal.cpp
@@ -8,6 +8,8 @@
 #if defined(OS_MAC) or defined(OS_LINUX)
 #	include <sys/ioctl.h>
 #	include <unistd.h>
+#elif defined(OS_WINDOWS)
+#include <windows.h>
 #endif
 
 namespace CppUtils::Terminal
@@ -57,8 +59,8 @@ namespace CppUtils::Terminal
 		auto consoleScreenBufferInfo = CONSOLE_SCREEN_BUFFER_INFO{};
 		GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &consoleScreenBufferInfo);
 		return Container::Size2d<>{
-			consoleScreenBufferInfo.srWindow.Right - consoleScreenBufferInfo.srWindow.Left + 1,
-			consoleScreenBufferInfo.srWindow.Bottom - consoleScreenBufferInfo.srWindow.Top + 1
+			static_cast<std::uint64_t>(consoleScreenBufferInfo.srWindow.Right - consoleScreenBufferInfo.srWindow.Left + 1),
+			static_cast<std::uint64_t>(consoleScreenBufferInfo.srWindow.Bottom - consoleScreenBufferInfo.srWindow.Top + 1)
 		};
 	}
 #elif defined(OS_MAC) or defined(OS_LINUX)
